refactor: colocate stack config and add list content semantics
