--- src/interface.ts
@@ -1,4 +1,0 @@
-export interface StackConfig {
-  threshold?: number;
-  offset?: number;
-}
